Move transaction descriptions in Description table
This way we reduce the number of transaction names.

I will be tying additiona information to the transaction names. For example: Costco at $100 will be categorized as groceries. A second trip to Costco at $150 will groceries again.

Now, if the next transaction is within that range, the probabily is very high that it's a grocery trip again.

If the trip is $60 and is marked as gas, the process repeates until more data points become available.